+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 **Bug fixes**
 
 - Fixed `onChange` typedef on `EuiSwitch` ([#1144](https://github.com/elastic/eui/pull/1144)
+- Fixed `EuiToolTip`'s inability to update its position when tooltip content changes ([#1116](https://github.com/elastic/eui/pull/1116))
 
 ## [`3.7.0`](https://github.com/elastic/eui/tree/v3.7.0)